Minimise the sweeper blocking serving.
Instead of locking and sweeping across the key map, track changes to
that map and pass the changes to the sweeper thread. That way, the
sweeper can process over the copy without obtaining the lock and
blocking serving. Processing over the copy generates a set of candidate
keys to check in the key map. Unless there's been a hash collision or a
non-empty table has been discarded, the sweep should generate no
candidates and not block the serving.