Modify GC to avoid building a reverse index.
Instead of building an offset->hash index, tag entries when they are
deleted and iterate/rehash entries in a GC. With lots of entries, the
limiting factor was the O(n log n) sort to build the reverse index.
Although keys are now rehashed on a GC, this works out to be faster in
practice until you get large keys (> ~256 bytes). It also makes a GC
O(n).